refactor(cli): extract run parser owner
